admin/requirements.txt
Django==3.1.3
djangorestframework==3.12.2
mysqlclient==2.0.1
django-mysql==3.9
django-cors-headers==3.5.0
pika==1.1.0
celery==4.4.7
redis
django-celery-beat==2.0.0
django-timezone-field==4.0
flower==0.9.7